requirements.txt
python-telegram-bot>=20.7
openai>=1.3.7
python-dotenv>=1.0.0
pydantic>=2.6.0
aiofiles>=23.2.1 